Fix some compiler warnings

+++ src/mpc-hc/AppSettings.cpp
@@ -1590,7 +1590,7 @@ void CAppSettings::LoadSettings()
     if (IsWindows10OrGreater()) {
         CRegKey key;
         if (ERROR_SUCCESS == key.Open(HKEY_CURRENT_USER, _T("Software\\Microsoft\\Windows\\CurrentVersion\\Themes\\Personalize"), KEY_READ)) {
-            DWORD useTheme = -1;
+            DWORD useTheme = (DWORD)-1;
             if (ERROR_SUCCESS == key.QueryDWORDValue(_T("AppsUseLightTheme"), useTheme)) {
                 if (0 == useTheme) {
                     bWindows10DarkThemeActive = true;
@@ -1598,7 +1598,7 @@ void CAppSettings::LoadSettings()
             }
         }
         if (ERROR_SUCCESS == key.Open(HKEY_CURRENT_USER, _T("Software\\Microsoft\\Windows\\DWM"), KEY_READ)) {
-            DWORD useColorPrevalence = -1;
+            DWORD useColorPrevalence = (DWORD)-1;
             if (ERROR_SUCCESS == key.QueryDWORDValue(_T("ColorPrevalence"), useColorPrevalence)) {
                 if (1 == useColorPrevalence) {
                     bWindows10AccentColorsEnabled = true;
@@ -2654,7 +2654,7 @@ void CAppSettings::CRecentFileListWithMoreInfo::WriteList() {
     auto pApp = AfxGetMyApp();
     pApp->WriteProfileString(m_section, nullptr, nullptr);
     int i = 1;
-    int m = rfe_array.GetCount() > m_maxSize ? m_maxSize : rfe_array.GetCount();
+    int m = rfe_array.GetCount() > m_maxSize ? m_maxSize : (int)rfe_array.GetCount();
     for (; i <= m; i++) {
         auto& r = rfe_array[i - 1];
         CString t;
